+++ packages/app-form-builder/src/admin/components/FormEditor/Tabs/EditTab/EditFieldDialog/GeneralTab.tsx
@@ -7,7 +7,7 @@ import { I18NInput } from "@webiny/app-i18n/admin/components";
 import { useI18N } from "@webiny/app-i18n/hooks/useI18N";
 import { validation } from "@webiny/validation";
 import { FbFormModelField } from "@webiny/app-form-builder/types";
-import {FormChildrenFunctionParams} from "@webiny/form/Form";
+import { FormChildrenFunctionParams } from "@webiny/form/Form";
 
 type GeneralTabProps = {
     field: FbFormModelField;