+++ packages/app-form-builder/src/editor/plugins/formSettings/components/TermsOfServiceSettings.tsx
@@ -3,7 +3,7 @@ import { Grid, Cell } from "@webiny/ui/Grid";
 import { I18NInput } from "@webiny/app-i18n/admin/components";
 import { Switch } from "@webiny/ui/Switch";
 import { get } from "lodash";
-import  { FormSettingsPluginRenderFunctionType } from "@webiny/app-form-builder/types";
+import { FormSettingsPluginRenderFunctionType } from "@webiny/app-form-builder/types";
 
 const TermsOfServiceSettings: FormSettingsPluginRenderFunctionType = ({ Bind, formData }) => {
     const enabled = get(formData, "termsOfServiceMessage.enabled");