+++ packages/app-page-builder/src/editor/plugins/elementSettings/align/HorizontalAlignAction.tsx
@@ -51,7 +51,6 @@ const HorizontalAlignAction = props => {
     return React.cloneElement(children, { onClick, icon: icons[align] });
 };
 
-export default connect<any, any, any>(
-    state => ({ element: getActiveElement(state) }),
-    { updateElement }
-)(HorizontalAlignAction);
+export default connect<any, any, any>(state => ({ element: getActiveElement(state) }), {
+    updateElement
+})(HorizontalAlignAction);