+++ packages/app-page-builder/src/editor/plugins/elementSettings/height/Settings.tsx
@@ -84,7 +84,6 @@ const Settings = props => {
     );
 };
 
-export default connect<any, any, any>(
-    state => ({ element: getActiveElement(state) }),
-    { updateElement }
-)(Settings);
+export default connect<any, any, any>(state => ({ element: getActiveElement(state) }), {
+    updateElement
+})(Settings);